requirements.txt
altgraph==0.12
BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
bleach==1.4.1
blinker==1.4
bs4==0.0.1
cssselect==0.9.1
cssutils==1.0
Django==1.8.2
dominate==2.1.12
enum34==1.0.4
Flask==0.10.1
Flask-Bcrypt==0.6.2
Flask-Bootstrap==3.3.5.6
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.0
flask-sendgrid==0.1.0
Flask-Uploads==0.1.3
Flask-User==0.6.3
Flask-WTF==0.12
gnureadline==6.3.3
gunicorn==19.6.0
html5lib==0.999999
idna==2.0
img2pdf==0.2
ipaddress==1.0.7
ipython==3.2.1
itsdangerous==0.24
Jinja2==2.8
LEPL==5.1.3
libusb1==1.5.0
lxml==3.4.4
macholib==1.7
mailtools==2
MarkupSafe==0.23
modulegraph==0.12.1
nltk==3.0.3
numpy==1.9.2
passlib==1.6.4
pbr==1.9.1
Pillow==3.2.0
py-bcrypt==0.4
pyasn1==0.1.7
pycparser==2.14
pycrypto==2.6.1
pyenchant==1.6.6
python-bcrypt==0.3.1
python-docx==0.8.5
python-firebase==1.2
python-gflags==3.0.5
queuelib==1.2.2
requests==2.9.1
six==1.9.0
sendgrid==3.1.10
python-postmark>=0.4.3
SQLAlchemy==1.0.8
stevedore==1.12.0
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
Werkzeug==0.10.4
WTForms==2.0.2
XlsxWriter==0.7.3
